Add a simple endpoint to pass data on to an action
We can't post Hubspot webhooks directly to Github Actions, so we need a bit of
middleware to trigger an action. We could host all the code where the middleware
is hosted, but it's easier to see the output of an action in Github Actions, and
be alerted about any issues and diagnose them there, rather than all the code
running in a black box on Heroku or similar.