Get a spreadsheet by a specific spreadsheet ID
The spreadsheet may not always the the first one in a series, and fetching
spreadsheets by array key is error prone, as the sheets can often be moved by
users. It's safer to specify the ID of the sheet (seen in the URL of the sheet with
a `?gid=xxxxx` query string.